Fixed my last dumb idea
That idea of copying the resources to a Global "DialogueTreeDictionary" was dumb. We shouldn't be editing the resources.
Greetings are now also [DialogueBranches], and the displayed greeting message is chosen using the existing condition system for topics.

The downside side is that creating one-line messages(for stuff like checking out, idk, a bookshelf or a sign) doesnt feel as good as before. I might add a secondary [DialogueComponent] just to handle those cases because creating an entire branch to display "<- West Town | East Town ->" is kinda dumb